requirements.txt
asgiref==3.8.1
awscli==1.36.31
botocore==1.35.90
colorama==0.4.6
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
docutils==0.16
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
whitenoise==6.8.2
